+++ cvat-cli/src/cvat_cli/cli.py
@@ -10,8 +10,8 @@ from typing import Dict, List, Sequence, Tuple
 
 import tqdm
 from cvat_sdk import Client, models
+from cvat_sdk.core.types import ResourceType
 from cvat_sdk.helpers import TqdmProgressReporter
-from cvat_sdk.types import ResourceType
 
 
 class CLI: