+++ cvat-sdk/gen/generate.sh
@@ -1,15 +1,20 @@
 #!/bin/sh
 
+# Copyright (C) 2022 CVAT.ai Corporation
+#
+# SPDX-License-Identifier: MIT
+
 set -e
 
 VERSION="2.0-alpha"
 LIB_NAME="cvat_sdk"
+LAYER1_LIB_NAME="${LIB_NAME}/api_client"
 DST_DIR="."
 TEMPLATE_DIR="gen"
 PYTHON_POST_PROCESS_FILE="${TEMPLATE_DIR}/postprocess.py"
 
 mkdir -p "${DST_DIR}/"
-rm -f -r "${DST_DIR}/docs/" "${DST_DIR}/${LIB_NAME}/api" "${DST_DIR}/${LIB_NAME}/model"
+rm -f -r "${DST_DIR}/docs/" "${DST_DIR}/${LAYER1_LIB_NAME}/api" "${DST_DIR}/${LAYER1_LIB_NAME}/model" "requirements/"
 cp "${TEMPLATE_DIR}/templates/openapi-generator/.openapi-generator-ignore" "${DST_DIR}/"
 
 # Pass template dir here
@@ -26,8 +31,9 @@ sudo chown -R "$(id -u)":"$(id -g)" "${DST_DIR}/"
 sed -e "s|{{packageVersion}}|${VERSION}|g" "${TEMPLATE_DIR}/templates/version.py.template" > "${DST_DIR}/${LIB_NAME}/version.py"
 cp -r "${TEMPLATE_DIR}/templates/requirements/" "${DST_DIR}/"
 cp -r "${TEMPLATE_DIR}/templates/MANIFEST.in" "${DST_DIR}/"
+mv "${DST_DIR}/requirements.txt" "${DST_DIR}/requirements/api_client.txt"
 
 # Do custom postprocessing
 "${PYTHON_POST_PROCESS_FILE}" --schema "schema/schema.yml" --input-path "${DST_DIR}/${LIB_NAME}"
-black --config pyproject.toml "${DST_DIR}/${LIB_NAME}"
-isort --sp pyproject.toml "${DST_DIR}/${LIB_NAME}"
+black --config pyproject.toml .
+isort --sp pyproject.toml .