+++ cvat-sdk/gen/postprocess.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+
 # Copyright (C) 2022 Intel Corporation
 #
 # SPDX-License-Identifier: MIT